server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
python-dotenv
pydantic-settings
httpx
python-jose[cryptography]
pydantic